remove non abstract abstract leftover classes, bugfixes
